Changes to imagemagick

+++ backend/src/data/image.ts
@@ -31,9 +31,12 @@ export class ImadeDataAccess {
           if (err) throw err;
         
           for (const file of files) {
-            fs.unlink(path.join(directory, file), err => {
-              if (err) throw err;
-            });
+            
+            if(file !== '.gitignore'){
+              fs.unlink(path.join(directory, file), err => {
+                if (err) throw err;
+              });
+            }       
           }
         });
     }
